+++ src/main/java/kr/or/kosa/snippets/like/mapper/SnippetContentMapper.java
@@ -10,11 +10,11 @@ import org.apache.ibatis.annotations.Param;
 public interface SnippetContentMapper {
 
     // 코드 스니펫 내용 조회
-    SnippetCode getSnippetCodeById(@Param("snippetId") Integer snippetId);
+    SnippetCode getSnippetCodeById(@Param("snippetId") Long snippetId);
 
     // 텍스트 스니펫 내용 조회
-    SnippetText getSnippetTextById(@Param("snippetId") Integer snippetId);
+    SnippetText getSnippetTextById(@Param("snippetId") Long snippetId);
 
     // 이미지 스니펫 내용 조회
-    SnippetImage getSnippetImageById(@Param("snippetId") Integer snippetId);
+    SnippetImage getSnippetImageById(@Param("snippetId") Long snippetId);
 }
